helm: add option to enable automatic etcd service name resolution
[ upstream commit 13533a2e96c4a90208446cac61ec835c65b54c90 ]

This feature is useful if users want to deploy an k8s service
that translates to an etcd cluster. As already implemented in Cilium, it
will perform the service translation internally without having to reach
a DNS to perform this translation. This will work by relying on the
watching mechanism that we have for k8s for which it is required that
`disabled-k8s-services` is set to `false`

Signed-off-by: André Martins <andre@cilium.io>
Signed-off-by: John Fastabend <john.fastabend@gmail.com>