Endpoint for host
This commit adds a special Cilium endpoint on each node to represent the
node itself. It takes a reserved endpoint ID 1 and the reserved security
ID 1. The special endpoint is created by endpoint.CreateLocalNodeEndpoint
on daemon startup if that endpoint wasn't already restored.

When the datapath for the special endpoint is regenerated, it loads
bpf_netdev.o instead of bpf_lxc.o. bpf_netdev enforces policies for both
traffic to/from pods and traffic to/from the outside world. It follows
the same template generation path as traditional bpf_lxc files.

Signed-off-by: Paul Chaignon <paul@cilium.io>