bpf,service: Add provisioning of sessionAffinity svc
First, this commit extends the lb{4,6}_service struct by allowing to
store the fact that sessionAffinity is enabled and a sessionAffinity
timeout (in seconds). The latter is stored only in the placeholder
svc value (slave=0), and it is stored in the backend_id field (thus
the change to the union type).

Second, it adds the provisioning of the sessionAffinity field from k8s
to the BPF maps.

Signed-off-by: Martynas Pumputis <m@lambda.lt>