WIP: test: Fix fragment tracking test under KUBEPROXY=1
The fragment tracking test currently fails when kube-proxy is enabled
because the source IP addresses are sometimes SNATed and the destination
IP address and port are sometimes not DNATed. The awk filter on the
ctmap dump output then fails.

This commit fixes those issues, but the test becomes more convoluted as
a result. In addition, the 'fromNode == helpers.K8s2' isn't handled
anymore. Maybe matching on the source port without the source IP address
would be enough?

Fixes: #10929
Cc: Quentin Monnet <quentin@isovalent.com>
Signed-off-by: Paul Chaignon <paul@cilium.io>